re-organize s3 download code
