+++ aws_s3_log_analyzer/s3.py
@@ -1,31 +1,11 @@
 import boto3
 import os
 
-from datetime import date as ddate
-
-def date1_since_date2(date1, date2):
-  ''' returns true if date1 is same as date2 or after, or if date2 is not a datetime object'''
-  if not isinstance(date2, ddate):
-    return True
-  else:
-    return date1 >= date2
-
-def date1_within_date2(date1, date2):
-  ''' returns true if date1 is same as date2 or before, or if date2 is not a datetime object '''
-  if not isinstance(date2, ddate):
-    return True
-  else:
-    return date1 <= date2
-
-def in_date_range(list1, list1_dates, start_date, end_date):
-  ''' return list whose elements are within date range '''
-  agrees_with_start_date = [x for idx, x in enumerate(list1) if date1_since_date2(list1_dates[idx], start_date)]
-  agrees_with_end_date = [x for idx, x in enumerate(list1) if date1_within_date2(list1_dates[idx], end_date)]
-  return [x for x in agrees_with_start_date if x in agrees_with_end_date]
+from aws_s3_log_analyzer.dates import in_date_range
 
 # adapted from: https://stackoverflow.com/questions/31918960/boto3-to-download-all-files-from-a-s3-bucket
 def download(bucket, start_date, end_date, dest_dir):
-  ''' replicate the content of S3 bucket to local dir '''
+  ''' copy the content of S3 bucket to local dir '''
   client = boto3.client('s3')
   keys = []
   keys_last_modified = []
